Modified raceleague page, created methods to read ballot data from db
