+++ server/Entities/BallotContent.cs
@@ -2,7 +2,6 @@
 
 public class BallotContent
 {
-    public required int BallotId { get; set; }
     public required int Position { get; set; }
     public required int DriverId { get; set; }
 }
